refactor: remove default exports
