+++ packages/common/dev/lib/main.ts
@@ -1,1 +1,1 @@
-export {default as matchers} from './matchers';
+export {matchers} from './matchers';
