+++ packages/common/dev/lib/matchers/index.ts
@@ -1,5 +1,5 @@
 import {toHaveFocusWithin} from './toHaveFocusWithin';
 
-export default {
+export const matchers = {
 	toHaveFocusWithin,
 };
