+++ packages/common/react/lib/main.ts
@@ -1,4 +1,4 @@
-export {default as useLayoutPromise} from './useLayoutPromise';
-export {default as useMultipleRefs} from './useMultipleRefs';
-export {default as useRunOnce} from './useRunOnce';
-export {default as useSyncedOption} from './useSyncedOption';
+export {useLayoutPromise} from './useLayoutPromise';
+export {useMultipleRefs} from './useMultipleRefs';
+export {useRunOnce} from './useRunOnce';
+export {useSyncedOption} from './useSyncedOption';
