+++ packages/common/svelte/lib/main.ts
@@ -1,3 +1,3 @@
-export {default as createEventForwarder} from './createEventForwarder';
-export {default as createSyncedOption} from './createSyncedOption';
+export {createEventForwarder} from './createEventForwarder';
+export {createSyncedOption} from './createSyncedOption';
 export * from './store';
