+++ packages/dialog/react/lib/main.ts
@@ -14,4 +14,4 @@ export const Dialog = {
 	Trigger,
 };
 
-export {default as useDialog} from './useDialog';
+export {useDialog} from './useDialog';
