+++ packages/dialog/vue/lib/main.ts
@@ -4,4 +4,4 @@ export {default as DialogDescription} from './DialogDescription.vue';
 export {default as DialogRoot} from './DialogRoot.vue';
 export {default as DialogTitle} from './DialogTitle.vue';
 export {default as DialogTrigger} from './DialogTrigger.vue';
-export {default as useDialog} from './useDialog';
+export {useDialog} from './useDialog';
